Implement more dockerflow requirements
